api/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.1
contextlib2==21.6.0
cryptography==44.0.0
Django==5.1.5
django-ninja==1.3.0
django-ninja-extra==0.22.2
django-ninja-jwt==5.3.5
dnspython==2.7.0
email_validator==2.2.0
fire==0.7.0
idna==3.10
injector==0.22.0
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
rav==0.0.9
rich==13.9.4
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.12.2
